Add ZoomChanged listener to MessageDialogue
When a zoom change occurs, the OS automatically scales the window based
on the scale factor. However, since fonts do not always scale linearly,
this can cause text in dialogues to be cut off.

This change adds a ZoomChanged listener to MessageDialogue. When a
zoomChanged event is triggered, the shell size is recomputed so that the
contents are fit properly and text is not clipped.